reorganize maputils
